REFPLTB-1671: USB flash: modify partition layout
There are two ways to flash the Turris:
- using USB drive
- using TurrisFwUpgrade.sh script

However they resulted in different outcome:
- different /etc/fstab
- nvram mounted on either:
  - /dev/mmcblk0p5 as a result of fstab entry
  - /dev/mmcblk0p6 mounted by hostapd-init.sh

Unify and optimize the flashing outcome.

As a first step modify the partition layout when flashing from USB:
- Remove nvram2 partition which is not used anyway, use the extra space
  to increase the size of other partitions
- Don't update the /etc/fstab when creating the wic image, which is used
  when flashing from USB
- Use --fixed-size option, so that the build fails, if the rootfs ever
  grows bigger than partitions allow

The USB flash method needs to be performed at least once on every unit
so that the right partition layout is established. Then script method
can be used.

Signed-off-by: Piotr Nakraszewicz <piotr.nakraszewicz@consult.red>
Signed-off-by: Piotr Nakraszewicz <pnakraszewicz@plume.com>