REFPLTB-1671: Turris extender flashing modifications (#451)
Add TurrisFwUpgrade.sh on the extender file system. Move it to a new
package - turris-flash - as sysint-broadband package is not a part of
extender image.

Use the same partition layout on extender as on gateway. This is
achieved by deleting the extender .wks file and renaming the gateway one
to a generic turris.wks

Signed-off-by: Piotr Nakraszewicz <piotr.nakraszewicz@consult.red>
Signed-off-by: Piotr Nakraszewicz <pnakraszewicz@plume.com>